example/requirements.txt
django
django-filter>0.15
django-currencies
